+++ src/math.ts
@@ -1,0 +1,3 @@
+export { GPUArray, init } from "./index";
+export type { AccelContext, ArrowImportOptions, ProfilingEntry } from "./types";
+export { add, sub, mul, div, sum, max, min, mean } from "./ops/math";